# k8s-client-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
        "/pod/createPod": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CreatePod"
                ],
                "summary": "获取单个 Pod 信息",
                "parameters": [
                    {
                        "description": "Pod simple configuration",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.PodReques"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/pod/deletePod": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DeletePod"
                ],
                "summary": "删除单个 Pod",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pod名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/pod/getAllPod": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetAllPod"
                ],
                "summary": "获取所有 Pod 信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/pod/getNamespacePod/{namespace}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetNamespacePod"
                ],
                "summary": "获取单个命名空间中所有的  Pod 信息",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        },
        "/pod/getPod": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetPod"
                ],
                "summary": "获取单个 Pod 信息",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pod名称",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CommonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.PodReques": {
            "type": "object",
            "properties": {
                "containerName": {
                    "type": "string"
                },
                "containerPort": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "podName": {
                    "type": "string"
                }
            }
        },
        "response.CommonResponse": {
            "type": "object",
            "properties": {
                "response message": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}